arrabbiata::challenge

Struct Challenges

pub struct Challenges<F> {
    pub constraint_combiner: F,
    pub beta: F,
    pub gamma: F,
    pub constraint_homogeniser: F,
    pub relation_combiner: F,
}

Fields
constraint_combiner: F

Used to aggregate the constraints describing the relation. It is used to enforce all constraints are satisfied at the same time. Often noted α.

beta: F

Both challenges used in the permutation argument.

gamma: Fconstraint_homogeniser: F

Used to homogenize the constraints and allow the protocol to fold two instances of the same relation into a new one. Often noted u in the paper mentioning “folding protocols”.

relation_combiner: F

Used by the accumulation protocol (folding) to perform a random linear transformation of the witnesses and the public values. Often noted r in the paper mentioning “folding protocols”.