arrabbiata::constraint

Struct Env

pub struct Env<C: ArrabbiataCurve>
where
    C::BaseField: PrimeField,
{
    pub a: BigInt,
    pub idx_var: usize,
    pub idx_var_next_row: usize,
    pub constraints: Vec<E<C::ScalarField>>,
}

Fields
a: BigInt

The parameter a is the coefficients of the elliptic curve in affine coordinates.

idx_var: usizeidx_var_next_row: usizeconstraints: Vec<E<C::ScalarField>>
Implementations

impl<C: ArrabbiataCurve> Env<C>
where C::BaseField: PrimeField,

pub fn new() -> Self

impl<C: ArrabbiataCurve> Env<C>
where C::BaseField: PrimeField,

pub fn get_all_constraints_for_verifier(&self) -> Vec<E<C::ScalarField>>

Get all the constraints for the verifier circuit, only.

The following gadgets are used in the verifier circuit:

Instruction::PoseidonFullRound and Instruction::PoseidonSpongeAbsorb to verify the challenges and the public IO.

Instruction::EllipticCurveScaling and Instruction::EllipticCurveAddition to accumulate the commitments

pub fn get_all_constraints(&self) -> Vec<E<C::ScalarField>>

Get all the constraints for the verifier circuit and the application.

pub fn get_all_constraints_indexed_by_gadget( &self, ) -> HashMap<Gadget, Vec<E<C::ScalarField>>>